crm-v2-backend/school_crm/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==3.0.0
email-validator==2.0.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
